Python: Revert Trampoline change
