+++ modules/python3/src/processortrampoline.cpp
@@ -26,35 +26,36 @@
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  *
  *********************************************************************************/
-#pragma once
+
+#include <modules/python3/processortrampoline.h>
 
 #include <warn/push>
 #include <warn/ignore/shadow>
 #include <pybind11/pybind11.h>
 #include <warn/pop>
 
-#include <inviwo/core/processors/processor.h>
-
 namespace inviwo {
 
-/**
- * A pybind11 trampoline class for Processor, needed to route virtual calls to/from python
- */
-class ProcessorTrampoline : public Processor, public pybind11::trampoline_self_life_support {
-public:
-    /* Inherit the constructors */
-    using Processor::Processor;
+/* Trampoline (need one for each virtual function) */
+void ProcessorTrampoline::initializeResources() {
+    PYBIND11_OVERLOAD(void, Processor, initializeResources, );
+}
+void ProcessorTrampoline::process() { PYBIND11_OVERLOAD(void, Processor, process, ); }
+void ProcessorTrampoline::doIfNotReady() { PYBIND11_OVERLOAD(void, Processor, doIfNotReady, ); }
+void ProcessorTrampoline::setValid() { PYBIND11_OVERLOAD(void, Processor, setValid, ); }
+void ProcessorTrampoline::invalidate(InvalidationLevel invalidationLevel,
+                                     Property* modifiedProperty) {
+    PYBIND11_OVERLOAD(void, Processor, invalidate, invalidationLevel, modifiedProperty);
+}
+const ProcessorInfo ProcessorTrampoline::getProcessorInfo() const {
+    PYBIND11_OVERLOAD_PURE(const ProcessorInfo, Processor, getProcessorInfo, );
+}
 
-    /* Trampoline (need one for each virtual function) */
-    virtual void initializeResources() override;
-    virtual void process() override;
-    virtual void doIfNotReady() override;
-    virtual void setValid() override;
-    virtual void invalidate(InvalidationLevel invalidationLevel,
-                            Property* modifiedProperty = nullptr) override;
-    virtual const ProcessorInfo getProcessorInfo() const override;
-    virtual void invokeEvent(Event* event) override;
-    virtual void propagateEvent(Event* event, Outport* source) override;
-};
+void ProcessorTrampoline::invokeEvent(Event* event) {
+    PYBIND11_OVERLOAD(void, Processor, invokeEvent, event);
+}
+void ProcessorTrampoline::propagateEvent(Event* event, Outport* source) {
+    PYBIND11_OVERLOAD(void, Processor, propagateEvent, event, source);
+}
 
 }  // namespace inviwo
